Fix timestep dtype: pass bfloat16 instead of long to HunyuanVideo15TimeEmbedding
HunyuanVideo15TimeEmbedding.forward() casts sinusoidal projections back
to timestep.dtype before the linear layer; passing torch.long caused a
dtype mismatch with bfloat16 weights.

+++ hunyuan_video_1_5_i2v_720p_gguf/pytorch/loader.py
@@ -296,7 +296,7 @@ class ModelLoader(ForgeModel):
             config.image_embed_dim,
             dtype=dtype,
         )
-        timestep = torch.tensor([500], dtype=torch.long).expand(batch_size)
+        timestep = torch.tensor([500], dtype=dtype).expand(batch_size)
 
         return {
             "hidden_states": hidden_states,
